Add server-rendered password generator shell
